[FE/#66] Feat : UserMenu 컴포넌트 분리
