+++ FE/src/App.js
@@ -1,7 +1,7 @@
 import React from "react";
 
 import GlobalStyle from "@Styles/GlobalStyle";
-import Header from "@Components/common/Header";
+import Header from "@Components/common/Header/Header";
 import Issues from "@Pages/Issues";
 import Container from "@material-ui/core/Container";
